+++ app/src/main/java/org/breezyweather/remoteviews/presenters/ClockDayDetailsWidgetIMP.kt
@@ -27,10 +27,10 @@ import breezyweather.domain.location.model.Location
 import breezyweather.domain.weather.model.Weather
 import org.breezyweather.R
 import org.breezyweather.background.receiver.widget.WidgetClockDayDetailsProvider
+import org.breezyweather.common.basic.models.options.appearance.CalendarHelper
 import org.breezyweather.common.extensions.currentLocale
 import org.breezyweather.common.extensions.getFormattedMediumDayAndMonthInAdditionalCalendar
 import org.breezyweather.common.extensions.getShortWeekdayDayMonth
-import org.breezyweather.common.extensions.isChinese
 import org.breezyweather.domain.location.model.getPlace
 import org.breezyweather.domain.location.model.isDaylight
 import org.breezyweather.domain.weather.model.getIndex
@@ -50,7 +50,7 @@ object ClockDayDetailsWidgetIMP : AbstractRemoteViewsPresenter() {
         val config = getWidgetConfig(context, context.getString(R.string.sp_widget_clock_day_details_setting))
         val views = getRemoteViews(
             context, location,
-            config.cardStyle, config.cardAlpha, config.textColor, config.textSize, config.clockFont, config.hideLunar
+            config.cardStyle, config.cardAlpha, config.textColor, config.textSize, config.clockFont, config.hideAlternateCalendar
         )
         AppWidgetManager.getInstance(context).updateAppWidget(
             ComponentName(context, WidgetClockDayDetailsProvider::class.java),
@@ -60,7 +60,7 @@ object ClockDayDetailsWidgetIMP : AbstractRemoteViewsPresenter() {
 
     fun getRemoteViews(
         context: Context, location: Location?,
-        cardStyle: String?, cardAlpha: Int, textColor: String?, textSize: Int, clockFont: String?, hideLunar: Boolean
+        cardStyle: String?, cardAlpha: Int, textColor: String?, textSize: Int, clockFont: String?, hideAlternateCalendar: Boolean
     ): RemoteViews {
         val color = WidgetColor(context, cardStyle!!, textColor!!, location?.isDaylight ?: true)
         val views = RemoteViews(
@@ -134,8 +134,10 @@ object ClockDayDetailsWidgetIMP : AbstractRemoteViewsPresenter() {
             )
         } ?: views.setViewVisibility(R.id.widget_clock_day_icon, View.INVISIBLE)
         views.setTextViewText(
-            R.id.widget_clock_day_lunar,
-            if (context.currentLocale.isChinese && !hideLunar) " - " + Date().getFormattedMediumDayAndMonthInAdditionalCalendar(location, context) else ""
+            R.id.widget_clock_day_alternate_calendar,
+            if (CalendarHelper.getAlternateCalendarSetting(context) != null && !hideAlternateCalendar) {
+                " - " + Date().getFormattedMediumDayAndMonthInAdditionalCalendar(location, context)
+            } else ""
         )
         val builder = StringBuilder()
         builder.append(location.getPlace(context))
@@ -177,7 +179,7 @@ object ClockDayDetailsWidgetIMP : AbstractRemoteViewsPresenter() {
                 setTextColor(R.id.widget_clock_day_clock_aa_normal, color.textColor)
                 setTextColor(R.id.widget_clock_day_clock_aa_black, color.textColor)
                 setTextColor(R.id.widget_clock_day_title, color.textColor)
-                setTextColor(R.id.widget_clock_day_lunar, color.textColor)
+                setTextColor(R.id.widget_clock_day_alternate_calendar, color.textColor)
                 setTextColor(R.id.widget_clock_day_subtitle, color.textColor)
                 setTextColor(R.id.widget_clock_day_todayTemp, color.textColor)
                 setTextColor(R.id.widget_clock_day_feelsLikeTemp, color.textColor)
@@ -200,7 +202,7 @@ object ClockDayDetailsWidgetIMP : AbstractRemoteViewsPresenter() {
                 setTextViewTextSize(R.id.widget_clock_day_clock_aa_normal, TypedValue.COMPLEX_UNIT_PX, clockAASize)
                 setTextViewTextSize(R.id.widget_clock_day_clock_aa_black, TypedValue.COMPLEX_UNIT_PX, clockAASize)
                 setTextViewTextSize(R.id.widget_clock_day_title, TypedValue.COMPLEX_UNIT_PX, contentSize)
-                setTextViewTextSize(R.id.widget_clock_day_lunar, TypedValue.COMPLEX_UNIT_PX, contentSize)
+                setTextViewTextSize(R.id.widget_clock_day_alternate_calendar, TypedValue.COMPLEX_UNIT_PX, contentSize)
                 setTextViewTextSize(R.id.widget_clock_day_subtitle, TypedValue.COMPLEX_UNIT_PX, contentSize)
                 setTextViewTextSize(R.id.widget_clock_day_todayTemp, TypedValue.COMPLEX_UNIT_PX, contentSize)
                 setTextViewTextSize(R.id.widget_clock_day_feelsLikeTemp, TypedValue.COMPLEX_UNIT_PX, contentSize)