+++ app/src/main/java/org/breezyweather/remoteviews/presenters/notification/MultiCityWidgetNotificationIMP.kt
@@ -26,10 +26,9 @@ import androidx.core.graphics.drawable.IconCompat
 import breezyweather.domain.location.model.Location
 import org.breezyweather.R
 import org.breezyweather.common.basic.models.options.NotificationTextColor
+import org.breezyweather.common.basic.models.options.appearance.CalendarHelper
 import org.breezyweather.common.basic.models.options.unit.TemperatureUnit
-import org.breezyweather.common.extensions.currentLocale
 import org.breezyweather.common.extensions.getFormattedMediumDayAndMonthInAdditionalCalendar
-import org.breezyweather.common.extensions.isChinese
 import org.breezyweather.common.extensions.notificationBuilder
 import org.breezyweather.common.extensions.notify
 import org.breezyweather.domain.location.model.getPlace
@@ -134,7 +133,7 @@ object MultiCityWidgetNotificationIMP : AbstractRemoteViewsPresenter() {
         } else current.temperature?.temperature
         val timeStr = StringBuilder()
         timeStr.append(location.getPlace(context))
-        if (context.currentLocale.isChinese) {
+        if (CalendarHelper.getAlternateCalendarSetting(context) != null) {
             timeStr.append(context.getString(R.string.comma_separator))
                 .append(Date().getFormattedMediumDayAndMonthInAdditionalCalendar(location, context))
         }